requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1

absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.14.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.52.4
gast==0.4.0
gdown==5.2.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.0
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
imutils==0.5.4
itsdangerous==2.2.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.8
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
rsa==4.9
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
soupsieve==2.5
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.14.1
